bdf: fix host contamination
WARNING: QA Issue: the-backdoor-factory: /the-backdoor-factory/usr/share/the-backdoor-factory/arm/LinuxARMLELF32.py is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination
the-backdoor-factory: /the-backdoor-factory/usr/share/the-backdoor-factory/arm/__init__.py is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination

Signed-off-by: Armin Kuster <akuster808@gmail.com>