dnsreson: fix host contamination
WARNING: QA Issue: dnsrecon: /dnsrecon/opt/dnsrecon/README.md is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination
dnsrecon: /dnsrecon/opt/dnsrecon/subdomains-top1mil-5000.txt is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination
dnsrecon: /dnsrecon/opt/dnsrecon/snoop.txt is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination

Signed-off-by: Armin Kuster <akuster808@gmail.com>