theHarvester: fix host contamination
WARNING: QA Issue: theHarvester: /theHarvester/opt/theHarvester/COPYING is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination
theHarvester: /theHarvester/opt/theHarvester/myparser.py is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination

Signed-off-by: Armin Kuster <akuster808@gmail.com>