discover: fix host contaminiation
WARNING: QA Issue: discover: /discover/opt/discover/README.md is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination
discover: /discover/opt/discover/crack-wifi.sh is owned by uid 1000, which is the same as the user running bitbake. This may be due to host contamination

Signed-off-by: Armin Kuster <akuster808@gmail.com>